# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minutor API",
        "description": "Sistema de apontamento de horas e despesas",
        "contact": {
            "email": "admin@minutor.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Listar customers",
                "description": "Lista customers com paginação, filtros e ordenação",
                "operationId": "a29d5f0bab561196b8e3bd82422e58e6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página (padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Registros por página (padrão: 15, máximo: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por name ou CGC",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "João"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordenação (ex: name,-created_at)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "name,-created_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hasNext": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "cgc": {
                                                        "type": "string"
                                                    },
                                                    "formatted_cgc": {
                                                        "type": "string"
                                                    },
                                                    "cgc_type": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Criar customer",
                "description": "Cria um novo customer no sistema",
                "operationId": "1d27da36718a3e9726c873633af496da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cgc"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do cliente",
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "cgc": {
                                        "description": "CPF ou CNPJ (apenas números)",
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "cgc": {
                                            "type": "string"
                                        },
                                        "formatted_cgc": {
                                            "type": "string"
                                        },
                                        "cgc_type": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "detailMessage": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido através do endpoint de login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}